test(input): migrate highlight
