+++ core/src/components/input/test/highlight/input.e2e.ts-snapshots/input-no-fill-focus-ios-ltr-Mobile-Firefox-linux.png
