+++ core/src/components/input/test/highlight/input.e2e.ts-snapshots/input-no-fill-focus-md-ltr-Mobile-Safari-linux.png
