+++ core/src/components/input/test/highlight/input.e2e.ts-snapshots/input-no-fill-invalid-ios-ltr-Mobile-Firefox-linux.png
