+++ core/src/components/input/test/highlight/input.e2e.ts-snapshots/input-no-fill-valid-ios-ltr-Mobile-Firefox-linux.png
