+++ core/src/components/input/test/highlight/input.e2e.ts-snapshots/input-outline-invalid-ios-ltr-Mobile-Chrome-linux.png
