+++ core/src/components/input/test/highlight/input.e2e.ts-snapshots/input-outline-invalid-md-ltr-Mobile-Safari-linux.png
