+++ core/src/components/input/test/highlight/input.e2e.ts-snapshots/input-outline-valid-md-ltr-Mobile-Chrome-linux.png
