+++ core/src/components/input/test/highlight/input.e2e.ts-snapshots/input-solid-focus-md-ltr-Mobile-Firefox-linux.png
