+++ core/src/components/input/test/highlight/input.e2e.ts-snapshots/input-solid-invalid-ios-ltr-Mobile-Safari-linux.png
