+++ core/src/components/input/test/highlight/input.e2e.ts-snapshots/input-solid-invalid-md-ltr-Mobile-Firefox-linux.png
