+++ core/src/components/input/test/highlight/input.e2e.ts-snapshots/input-solid-valid-md-ltr-Mobile-Firefox-linux.png
